Update publishable stubs
